Add files via upload
This study was done to deduce whether two brain regions (Amygdala and Anterior Cingulate Cortex - ACC) are independent of each other and identify correlations of the residuals of the volumetric size of amygdala and acc with political orientation values (1- most conservative to 5 - most liberal).  

The result of this study was that higher values in the political orientation scale tend to have higher amygdala and acc volumes associated with them.
To identify the correlation, I utilized Kernel Density Estimation (KDE) to estimate the distributions of the amygdala and acc volumes in 2-Dimensions and identify the conditional distribution of the variables conditioned on political orientation.